Safari-609.1.11.1/LayoutTests/fast/inspector-support/uncaught-dom3-exception-expected.txt

CONSOLE MESSAGE: line 9: HierarchyRequestError: The operation would yield an incorrect node tree.
This tests that we are getting the correct message for DOM Exception 3: HIERARCHY_REQUEST_ERR.
